pyproject.toml
[tool.poetry]
name = "automl-paper"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.26.0"
scikit-learn = "^1.5.0"
scipy = "^1.13.1"
optuna = "^3.6.1"
catboost = "^1.2.5"
pandas = "^2.2.2"
descriptastorus = {git = "https://github.com/bp-kelley/descriptastorus"}
rdkit = "^2023.9.6"
pytdc = "^0.4.17"
networkx = "^3.3"
joblib = "^1.4.2"
mlflow = "^2.16.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.27.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
